# src/models/api/messageResponse.ts
/**
 * Sesamo
 * APIs for Sesamo project
 *
 * OpenAPI spec version: 0.0.1
 * Contact: matteopietro.dazzi@gft.com
 *
 * NOTE: This class is auto generated by the swagger code generator program.
 * https://github.com/swagger-api/swagger-codegen.git
 * Do not edit the class manually.
 */

export class MessageResponse {
    constructor(private messageId: string) {
    }
}